refactor: modernize codebase, fix localization issues, and optimize UI components
- Enforced `Locale.US` in all `String.format` calls to ensure consistent numeric formatting across different regions.
- Migrated `SharedPreferences` operations to use the KTX `edit` trailing lambda for cleaner syntax.
- Updated various Compose icons to their `AutoMirrored` versions to support right-to-left (RTL) layouts.
- Refactored URI parsing to use the `toUri()` KTX extension for better readability.
- Optimized Compose performance by remembering `MutableInteractionSource` and using `mutableIntStateOf`.
- Cleaned up redundant dependencies in `build.gradle.kts` and removed unused color resources.
- Improved `NavigationVoicePrompt` reliability with refined TTS locale initialization and fallback logic.
- Simplified Hilt `@ApplicationContext` injection syntax across multiple data stores and services.
- Refined `RideShareCard` rendering logic using `canvas.withSave` and improved path clipping.
- Removed manual status and navigation bar color overrides in `Theme.kt` to allow standard Material 3 handling.

+++ app/src/main/java/com/odys/mototriptracker/ui/tracker/LiveRideMapView.kt
@@ -130,5 +130,5 @@ private fun zoomFromDistance(distanceMeters: Double): Float {
         distanceMeters <= 1000 -> 16.2f
         distanceMeters <= 1400 -> 15.6f
         else -> 15f
-    }.toFloat()
+    }
 }
